Adressing comments
https://favro.com/organization/c84b360389a51a68251fb542/0be15a48533818256ef5af3b?card=int-594

Added and used a new functional colour (badgeBackgroundDark) to the pallete.
Removed useless css selectors for comment action menu button.
Removed old theme declaration from Comment styledComponent
Placed potential commentHeader is a separate div and added more specificity to it's selection for styling.
Pushing commentHeaderDate to a new line if there is a commentHeader present.
Renamed the DropdownMenu styling in the Comment declaration.