Use UnstyledLink over Anchor and make the latter internal-only
The issue with anchor at the moment is that it applies styles
to the a tag to remove the colour of the anchor. This causes
issues with any component expecting the anchor to be unstyled.

I think it makes much more sense for unstyled links to be
handled similar to text/action links, which apply style changes
to the default anchor component. This component is assumed to
have the default anchor styles (i.e., being blue). This way
lapis/SPA can just use standard <a> tags for anchors rather
than BaseAnchor and having a weird 4 way import between
the two repositories.
 (TextLink -> Anchor -> RoutedAnchor -> BaseAnchor)

INT-956