Use column flex layout instead of max width stacking
Since there is only one header component the sass can now be
heavily simplified due to the fact that there is no need for two items
to be on the same line. It also removes the need for an inner and outer
subcomponent which cleanes up the dom and makes legacy implementations
neater.

[INT-2237]