Create the XYGrid component based off foundation
This component still uses foundation's very complicated method
of generating a grid in sass. I wanted to copy it, but there was
too much work involved for the time being, and the important
part is that it functions in react for now.

This supports most of the XYGrid functionality except for
frame grids and block grids, which are basically unused anyway.

INT-2332