Upgrade styled-components to v4
innerRef prop is now just ref

StyledFunction accepts a function type instead of just the props

`as` prop introduced, it can be used to dynaically change the tag
of a styled component. Needs to be `undefined` if the props passed
extend a specific html element.

StyledComponentClass is now StyledComponentBase, all references
removed because they weren't being used anyway

[INT-6543]