Fixed styleguide github push
Something might have changed in how git worktrees work.

Before:
After the a new worktree was created from the latest remote branch, the script would run the styleguide, comit changes and push.

Issue:
After the styleguide would build, the reference about the worktree on the styleguide dir would ... disapear?

Fix:
Worktree is checked out in a separate dir
After the styleguide finishes the build, the contents of the build are copied into the worktree dir, comitted and pushed to origin.